Initial commit
I have created the basic files for the git repository, as well as
implemented a very simple plotter for SFML. This plotting tool can be
used in the mean time for developing the core functionality of the
simulator. In the future I hope to replace this plotter with one using
Qt.