Did major refactoring, started random algorithm
I added support for command line arguments, multiple modes (auto and
step-through), printing help to the command line by default, changing
settings on the command line, seeing the default settings and the
current available algorithms on the command line, and running different
algorithms. I implemented the beginnings of the random move mapping
algorithm, which should be the simplest algorithm.