Added no_backtrack_random algorithm
I added no_backtrack_random algorithm, which is like the random
algorithm except it cannot move past to previous positions. This makes
the algorithm more efficient when it works, but also means the algorithm
gets stuck most of the time. I might remove this algorithm in the
future. The program now counts the number of iterations for each
algorithm as well. In the future I hope to make a helper functions file
that provides many useful functions to all the algorithms. That way
there is less copy and paste between algorithms.