Added the fast_deterministic algorithm
I added an algorithm which is not dependent on anything random, but
which also should never fail to find all obstacles (assuming there are
no bugs in the implementation). How it works is a little complicated,
and I may document it later, but it is too complicated to explain in a
git commit. The important points are that it is non-optimal,
deterministic (the same output given the same obstacles), and should
work on all grid sizes.