Started separating logic code from UI code
In order to start the process of porting the project to Qt, I have moved
almost all SFML code into its own class called SFMLUI. I plan to make a
ConsoleUI class that will provide the option to run the program
exclusively from the console in the future, and a QtUI that will run the
Qt frontend. The ConsoleUI would be especially helpful for writing
automated tests since a testing program could read directly from stdout.